refactor: dispatcher provenance lives in its own module
Moves the create-side stamping and the send-side tab resolution out of
handler-helpers.ts / handlers-tasks.ts into cli/api/dispatcher.ts so the
loop's reply-address logic reads as one concern (and handler-helpers.ts
stays untouched by this change).
